Title: 진수 to 진수, Time: -1ms, Memory: -1MB, Status: Runtime Error - CodeTree

+++ 241009/진수 to 진수/transformation-of-number-system.java
@@ -11,21 +11,21 @@ public class Main {
         StringTokenizer st = new StringTokenizer(br.readLine());
         int a = Integer.parseInt(st.nextToken());
         int b = Integer.parseInt(st.nextToken());
-        int n = Integer.parseInt(br.readLine());
+        long n = Integer.parseInt(br.readLine());
 
         n = getNumberByBinary(n, a);
         n = getNumberToBinary(n, b);
         System.out.println(n);
     }
 
-    private static int getNumberToBinary(int n, int b) {
+    private static int getNumberToBinary(long n, int b) {
         // 10 진수인 n 을 b 진수로 바꾸기
         List<Integer> list = new ArrayList<>();
         while (n != 0) {
-            list.add(n%b);
+            list.add((int) (n%b));
             n /= b;
         }
-        
+
         StringBuilder sb = new StringBuilder();
         for(int i=list.size()-1; i>=0; i--) {
             sb.append(list.get(i));
@@ -33,14 +33,14 @@ public class Main {
         return Integer.parseInt(sb.toString());
     }
 
-    private static int getNumberByBinary(int n, int a) {
+    private static int getNumberByBinary(long n, int a) {
         // a 진수인 값 n 을 십진수로 바꾸기
         int newNum = 0;
         String num = String.valueOf(n);
         for(int i=0; i<num.length(); i++){
             newNum = (newNum * a) + Character.getNumericValue(num.charAt(i));
         }
-        
+
         return newNum;
     }
 }
